Remove fun, add write file, fix plot bug from plots_out.R
Removed validate_parsed_args().
A few files of interested are now written out as csvs. These are needed
for some stats about the sample distribution. This was done quickly and
needed to be reworked to use write_file_list like the rest of the
workflow.
There was a bug where the combined plot had a typo.